chore: restructure near and reorganize dependencies
