+++ packages/axelar-local-dev-near/package.json
@@ -7,15 +7,17 @@
         "contract-test": "npx hardhat test",
         "test": "jest",
         "build-ts": "rm -rf dist && tsc",
-        "build-near": "cd ./src/near/contracts/test/near-axelar-contract-call-example && ./build.sh",
-        "build-contract": "rm -rf src/artifacts && hardhat compile && mv artifacts src/artifacts",
-        "copy-near-wasm": "cp ../../node_modules/@axelar-network/axelar-cgp-near/dist/axelar_cgp_near.wasm src/near/contracts/ && mkdir -p dist/near/contracts && cp -r src/near/contracts/axelar_cgp_near.wasm dist/near/contracts",
-        "build": "run-s build-near build-contract build-ts copy-near-wasm",
+        "build-near": "cd ./src/contracts/test/near-axelar-contract-call-example && ./build.sh",
+        "build-contract": "rm -rf artifacts src/artifacts && hardhat clean && hardhat compile && mv artifacts src/artifacts",
+        "copy-near-wasm": "cp ../../node_modules/@axelar-network/axelar-cgp-near/dist/axelar_cgp_near.wasm src/contracts/ && mkdir -p dist/contracts && cp -r src/contracts/axelar_cgp_near.wasm dist/contracts",
+        "build": "run-s build-near build-contract",
+        "prepare": "run-s build",
         "lint": "solhint 'src/contracts/**/*.sol' && eslint 'test/**/*.js'",
         "prettier": "prettier --write 'src/contracts/**/*.sol' 'test/**/*.js' 'src/**/*.ts'"
     },
     "dependencies": {
-        "@axelar-network/axelar-local-dev": "*"
+        "@axelar-network/axelar-local-dev": "*",
+        "@axelar-network/axelar-cgp-near": "^1.0.0"
     },
     "author": "",
     "license": "ISC"
