+++ packages/axelar-local-dev-near/src/contracts/.gitignore
