+++ packages/axelar-local-dev-near/src/contracts/test/near-axelar-contract-call-example/README.md
