+++ packages/axelar-local-dev-near/src/index.ts
