+++ packages/axelar-local-dev-near/src/nearNetworkUtils.ts
