+++ packages/axelar-local-dev/package.json
@@ -7,9 +7,9 @@
         "contract-test": "npx hardhat test",
         "test": "jest",
         "build-ts": "rm -rf dist && tsc",
-        "build-contract": "rm -rf src/artifacts && hardhat compile && mv artifacts src/artifacts",
+        "build-contract": "rm -rf src/artifacts artifacts && hardhat clean && hardhat compile && mv artifacts src/artifacts",
         "build": "run-s build-contract build-ts",
-        "postinstall": "run-s build",
+        "prepare": "run-s build",
         "lint": "solhint 'src/contracts/**/*.sol' && eslint 'test/**/*.js'",
         "prettier": "prettier --write 'src/contracts/**/*.sol' 'test/**/*.js' 'src/**/*.ts'"
     },
@@ -31,8 +31,6 @@
         "dist"
     ],
     "dependencies": {
-        "@axelar-network/axelar-cgp-aptos": "^1.0.5",
-        "@axelar-network/axelar-cgp-near": "^1.0.0",
         "@axelar-network/axelar-cgp-solidity": "github:axelarnetwork/axelar-cgp-solidity#feat/forecaller-service",
         "@axelar-network/axelar-gmp-sdk-solidity": "^3.2.1",
         "ethers": "^5.6.5",
